move form hook out to be reusable
