+++ client/components/pages/submission-system/modals/addUser/index.js
@@ -5,122 +5,11 @@ import Button from 'uikit/Button';
 import Icon from 'uikit/Icon';
 import Typography from 'uikit/Typography';
 import css from '@emotion/css';
+
 import { UserSection } from '../styledComponents';
-import * as yup from 'yup';
-import { get, isArray } from 'lodash';
-import addUserSchema from './validations';
+import userSchema from '../validations';
 import { UserModel } from '../common';
-
-const useFormHook = ({ initialFields, schema: formSchema }) => {
-  const [form, setForm] = useState({ errors: [initialFields], data: [initialFields] });
-  const [touched, setTouched] = useState(false);
-  const { errors, data } = form;
-
-  const hasErrors = errors
-    .map(section => Object.values(section))
-    .flat()
-    .some(x => x);
-
-  // set form data
-  const setData = ({ key, val, index }) => {
-    if (!touched) setTouched(true);
-
-    setForm({
-      ...form,
-      data: data.map((field, i) => (i === index ? { ...field, [key]: val } : field)),
-    });
-  };
-
-  // set single error
-  const setError = ({ key, val, index }) => {
-    setForm({
-      ...form,
-      errors: errors.map((field, i) => (i === index ? { ...field, [key]: val } : field)),
-    });
-  };
-
-  // set all errors
-  const setErrors = ({ validationErrors, index }) =>
-    setForm({
-      ...form,
-      errors: errors.map((field, i) => (i === index ? validationErrors : field)),
-    });
-
-  // delete a block of values
-  const deleteSection = deletedIndex => {
-    const filterDeleted = (item, index) => index !== deletedIndex;
-    setForm({
-      errors: errors.filter(filterDeleted),
-      data: data.filter(filterDeleted),
-    });
-  };
-
-  // create a block of values
-  const createSection = sectionFields => {
-    setForm({
-      errors: errors.concat(sectionFields),
-      data: data.concat(sectionFields),
-    });
-  };
-
-  // validate a single field
-  const validateField = async ({ key, index }) => {
-    try {
-      const value = await yup.reach(formSchema, key).validate(data[index][key]);
-      setError({ key, val: '', index });
-    } catch (fieldError) {
-      const message =
-        isArray(fieldError.inner) && fieldError.inner.length > 1
-          ? fieldError.inner[fieldError.inner.length - 1].message
-          : fieldError.message;
-
-      setError({ key, val: message, index });
-    }
-  };
-
-  // validate a section
-  const validateSection = async ({ index }) =>
-    new Promise(async (resolve, reject) => {
-      const section = data[index];
-      try {
-        const validData = await addUserSchema.validate(section, {
-          abortEarly: false,
-          stripUnknown: true,
-        });
-        resolve(validData);
-      } catch (formErrors) {
-        const validationErrors = get(formErrors, 'inner', []).reduce((output, error) => {
-          output[error.path] = error.message;
-          return output;
-        }, {});
-
-        setErrors({ validationErrors, index });
-        reject(validationErrors);
-      }
-    });
-
-  // validates entire form
-  const validateForm = () =>
-    Promise.all(
-      data.map(
-        (section, index) =>
-          new Promise(async (resolve, reject) => await validateSection({ index })),
-      ),
-    );
-
-  return {
-    errors,
-    data,
-    setData,
-    deleteSection,
-    createSection,
-    validateField,
-    validateSection,
-    validateForm,
-    touched,
-    hasErrors,
-  };
-};
+import useFormHook from '../useFormHook';
 
 const AddSection = styled(Button)`
   text-transform: uppercase;
@@ -141,7 +30,7 @@ const AddUserModal = ({ dismissModal }) => {
     validateForm,
     touched,
     hasErrors,
-  } = useFormHook({ initialFields: UserModel, schema: addUserSchema });
+  } = useFormHook({ initialFields: UserModel, schema: userSchema });
 
   const islastSectionTouched = Object.values(form[form.length - 1]).reduce(
     (acc, val) => acc || !!val,
